+++ example/parse_args_example.dart
@@ -1,4 +1,8 @@
+// Copyright (c) 2022, Alexander Iurovetski
+// All rights reserved under MIT license (see LICENSE file)
+
 import 'dart:io';
+import 'package:path/path.dart' as p;
 import 'package:parse_args/parse_args.dart';
 
 /// Application options
@@ -6,7 +10,6 @@ import 'package:parse_args/parse_args.dart';
 class Options {
   static const appName = 'sampleapp';
   static const appVersion = '2.0.1';
-  static const optDefStr = '?,h,help|q,quiet|v,verbose||c,appconfig:|d,dir:|f,force|i,inp,input::|o,out,output::|p,compression:i';
 
   var _appConfigPath = '';
   get appConfigPath => _appConfigPath;
@@ -32,59 +35,90 @@ class Options {
   final _outputFiles = <String>[];
   get outputFiles => _outputFiles;
 
+  /// General-purpose method to add file paths to destinaltion list
+
+  void addPaths(List<String> to, List from) {
+    for (var x in from) {
+      to.add(p.isAbsolute(x) ? x : p.join(_startDirName, x));
+    }
+  }
+
   /// Sample application's command-line parser
 
   void parse(List<String> args) {
-    parseArgs(optDefStr, args, (optName, values) {
-      switch (optName) {
-        case '':
-          if (optName.isEmpty) {
-            printUsage('${Options.appName} does not support plain arguments');
-          }
-          break;
-        case '?':
-        case 'h':
-        case 'help':
-          printUsage();
-        case 'c':
-        case 'appconfig':
-          _appConfigPath = values[0];
-          break;
-        case 'd':
-        case 'dir':
-          _startDirName = values[0];
-          break;
-        case 'f':
-        case 'force':
-          _isForced = true;
-          break;
-        case 'i':
-        case 'inp':
-        case 'input':
-          _inputFiles.addAll(values as List<String>);
-          break;
-        case 'o':
-        case 'out':
-        case 'output':
-          _outputFiles.addAll(values as List<String>);
-          break;
-        case 'p':
-        case 'compression':
-          _compression = values[0];
-          break;
-        case 'q':
-        case 'quiet':
-          _isQuiet = true;
-          break;
-        case 'v':
-        case 'verbose':
-          _isVerbose = true;
-          break;
-        default:
-          break;
+    parseArgs('+|q,quiet|v,verbose|?,h,help|c,appconfig:|d,dir:|f,force|i,inp,input::|o,out,output::|p,compression:i',
+              args, (isFirstRun, optName, values) {
+      if (isFirstRun) {
+        switch (optName) {
+          case 'compression':
+            _compression = values[0];
+            return;
+          case 'help':
+            printUsage();
+          case 'dir':
+            _startDirName = values[0];
+            return;
+          case 'force':
+            _isForced = true;
+            break;
+          case 'quiet':
+            _isQuiet = true;
+            return;
+          case 'verbose':
+            _isVerbose = true;
+            return;
+          default:
+            return;
+        }
+      }
+      else {
+        printVerbose('Parsing $optName => $values');
+
+        // No need to assign any option value here if it does not depend on another option value
+        // In this case, just print the info
+
+        switch (optName) {
+          case 'appconfig':
+            _appConfigPath = p.join(_startDirName, values[0]);
+            printInfo('...appConfigPath: $_appConfigPath');
+            return;
+          case 'compression':
+            printInfo('...compression: $_compression');
+            return;
+          case 'dir':
+            printInfo('...startDirName: $_startDirName');
+            return;
+          case 'force':
+            printInfo('...isForced: $_isForced');
+            break;
+          case 'input':
+            addPaths(_inputFiles, values);
+            printInfo('...inp file(s): $_inputFiles');
+            return;
+          case 'output':
+            addPaths(_outputFiles, values);
+            printInfo('...out file(s): $_outputFiles');
+            return;
+          case 'quiet':
+            printInfo('...quiet: $_isQuiet');
+            return;
+          case 'verbose':
+            printInfo('...verbose: $_isVerbose');
+            return;
+          default:
+            return;
+        }
       }
     });
   }
+
+  /// A very simple info logger
+
+  void printInfo(String line) { if (!_isQuiet) { print(line); } }
+
+  /// A very simple verbose logger
+
+  void printVerbose(String line) { if (!_isQuiet && _isVerbose) { print(line); } }
 
   /// Displaying the help and optionally, an error message
 
@@ -127,7 +161,6 @@ void main(List<String> args) {
   try {
     var o = Options();
     o.parse(args);
-
     // the rest of processing
   } on OptException catch (e) {
     e.print();