+++ lib/parse_args.dart
@@ -1,10 +1,7 @@
-// Copyright (c) Alexander Iurovetski 2022
-// All rights reserved under MIT license
+// Copyright (c) 2022, Alexander Iurovetski
+// All rights reserved under MIT license (see LICENSE file)
 
 library parse_args;
 
-export 'src/opt_def.dart';
 export 'src/opt_exception.dart';
-export 'src/opt_value_mode.dart';
-export 'src/opt_value_type.dart';
 export 'src/parse_args_base.dart';