+++ lib/src/opt_exception.dart
@@ -1,70 +1,72 @@
+// Copyright (c) 2022, Alexander Iurovetski
+// All rights reserved under MIT license (see LICENSE file)
+
 import 'dart:io';
 
 ///
 /// Base exception class for command-line options parsing
 ///
 class OptException implements Exception {
-  String displayData = '';
-  String get displayType => '';
+  String get description => '';
 
-  final String optName;
+  String details = '';
+  final String name;
   final List? values;
 
-  OptException(this.optName, [this.values]) {
-    String displayData;
-
-    if (values == null) {
-      displayData = '';
-    } else if ((values?.length ?? 0) == 1) {
-      displayData = values?.first?.toString() ?? '';
-    } else {
-      displayData = values.toString();
-    }
-
-    displayData = '"$optName"$displayData';
+  OptException(this.name, [this.values]) {
+    var hasValue = (values?.isNotEmpty ?? false);
+    details = '"$name"${hasValue ? ': ${values.toString()}' : ''}';
   }
 
   @override String toString() {
-    return '$displayType - $displayData';
+    return '$description - $details';
   }
 
   void print() =>
     stderr.writeln(toString());
 }
 
+///
+/// When encountered an unrecognised option
+///
 class OptNameException extends OptException {
-  @override String get displayType => 'Invalid option';
+  @override String get description => 'Invalid option';
 
-  OptNameException(optName, [values]) : super(optName, values);
+  OptNameException(optName) : super(optName);
 }
 
-class OptValueExpectedException extends OptException {
-  @override String get displayType => 'Expected value';
-
-  OptValueExpectedException(optName, [values]) : super(optName, values);
-}
-
+///
+/// When encountered an option with no value
+///
 class OptValueMissingException extends OptException {
-  @override String get displayType => 'Missing value';
+  @override String get description => 'Missing value';
 
-  OptValueMissingException(optName, [values]) : super(optName, values);
+  OptValueMissingException(optName) : super(optName);
 }
 
+///
+/// When encountered an option with more than one value when only one expected
+///
 class OptValueTooManyException extends OptException {
-  @override String get displayType => 'Too many values';
+  @override String get description => 'Too many values';
 
   OptValueTooManyException(optName, [values]) : super(optName, values);
 }
 
+///
+/// When failed either to recognise the verb or to parse the string value
+///
 class OptValueTypeException extends OptException {
-  @override String get displayType => 'Bad value type';
+  @override String get description => 'Bad value type';
 
   OptValueTypeException(optName, [values]) : super(optName, values);
 }
 
+///
+/// When encountered an option with value(s) intead of a flag
+///
 class OptValueUnexpectedException extends OptException {
-  @override String get displayType => 'Value is not expected ';
+  @override String get description => 'Value is not expected ';
 
   OptValueUnexpectedException(optName, [values]) : super(optName, values);
 }
-