+++ lib/src/opt_value_mode.dart
@@ -1,5 +1,10 @@
+// Copyright (c) 2022, Alexander Iurovetski
+// All rights reserved under MIT license (see LICENSE file)
+
+/// Enumeration for the number of the option values
+
 enum OptValueMode {
-  none,
-  single,
-  multiple,
+  none, // no value
+  single, // one value
+  multiple, // one or more values
 }