+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0
+
+- Added ability to define options to parse as well as to automate options/values validations.
+- Added ability to get values pre-converted to one of the known types (binary, decimal, hexadecimal, octal, float, string).
+- Added ability to perform arguments parsing twice (options definition string is parsed once as before). This is useful when you need to know more than one value while parsing. For instance, if the start-in directory and a filename are both options, you can't assign the file path straightaway
+- Full compatibility with the older version - just pass the null or empty string.
+
 ## 0.1.1
 
 - Fixed Dart formatting issues.